# Simplify Parser and Introduce Formatter