--- a/Sources/XcbeautifyLib/JunitReporter.swift
+++ b/Sources/XcbeautifyLib/JunitReporter.swift
@@ -50,51 +50,51 @@ package final class JunitReporter {
     }
 
     private func generateFailingTest(line: String) -> TestCase? {
-        guard let _group: any CaptureGroup = line.captureGroup(with: FailingTestCaptureGroup.pattern) else { return nil }
-        guard let group = _group as? FailingTestCaptureGroup else { return nil }
+        let groups = FailingTestCaptureGroup.regex.captureGroups(for: line)
+        guard let group = FailingTestCaptureGroup(groups: groups) else { return nil }
         return TestCase(classname: group.testSuite, name: group.testCase, time: nil, failure: .init(message: "\(group.file) - \(group.reason)"))
     }
 
     private func generateRestartingTest(line: String) -> TestCase? {
-        guard let _group: any CaptureGroup = line.captureGroup(with: RestartingTestCaptureGroup.pattern) else { return nil }
-        guard let group = _group as? RestartingTestCaptureGroup else { return nil }
+        let groups = RestartingTestCaptureGroup.regex.captureGroups(for: line)
+        guard let group = RestartingTestCaptureGroup(groups: groups) else { return nil }
         return TestCase(classname: group.testSuite, name: group.testCase, time: nil, failure: .init(message: line))
     }
 
     private func generateParallelFailingTest(line: String) -> TestCase? {
         // Parallel tests do not provide meaningful failure messages
-        guard let _group: any CaptureGroup = line.captureGroup(with: ParallelTestCaseFailedCaptureGroup.pattern) else { return nil }
-        guard let group = _group as? ParallelTestCaseFailedCaptureGroup else { return nil }
+        let groups = ParallelTestCaseFailedCaptureGroup.regex.captureGroups(for: line)
+        guard let group = ParallelTestCaseFailedCaptureGroup(groups: groups) else { return nil }
         return TestCase(classname: group.suite, name: group.testCase, time: nil, failure: .init(message: "Parallel test failed"))
     }
 
     private func generatePassingTest(line: String) -> TestCase? {
-        guard let _group: any CaptureGroup = line.captureGroup(with: TestCasePassedCaptureGroup.pattern) else { return nil }
-        guard let group = _group as? TestCasePassedCaptureGroup else { return nil }
+        let groups = TestCasePassedCaptureGroup.regex.captureGroups(for: line)
+        guard let group = TestCasePassedCaptureGroup(groups: groups) else { return nil }
         return TestCase(classname: group.suite, name: group.testCase, time: group.time)
     }
 
     private func generateSkippedTest(line: String) -> TestCase? {
-        guard let _group: any CaptureGroup = line.captureGroup(with: TestCaseSkippedCaptureGroup.pattern) else { return nil }
-        guard let group = _group as? TestCaseSkippedCaptureGroup else { return nil }
+        let groups = TestCaseSkippedCaptureGroup.regex.captureGroups(for: line)
+        guard let group = TestCaseSkippedCaptureGroup(groups: groups) else { return nil }
         return TestCase(classname: group.suite, name: group.testCase, time: group.time, skipped: .init(message: nil))
     }
 
     private func generatePassingParallelTest(line: String) -> TestCase? {
-        guard let _group: any CaptureGroup = line.captureGroup(with: ParallelTestCasePassedCaptureGroup.pattern) else { return nil }
-        guard let group = _group as? ParallelTestCasePassedCaptureGroup else { return nil }
+        let groups = ParallelTestCasePassedCaptureGroup.regex.captureGroups(for: line)
+        guard let group = ParallelTestCasePassedCaptureGroup(groups: groups) else { return nil }
         return TestCase(classname: group.suite, name: group.testCase, time: group.time)
     }
 
     private func generateSkippedParallelTest(line: String) -> TestCase? {
-        guard let _group: any CaptureGroup = line.captureGroup(with: ParallelTestCaseSkippedCaptureGroup.pattern) else { return nil }
-        guard let group = _group as? ParallelTestCaseSkippedCaptureGroup else { return nil }
+        let groups = ParallelTestCaseSkippedCaptureGroup.regex.captureGroups(for: line)
+        guard let group = ParallelTestCaseSkippedCaptureGroup(groups: groups) else { return nil }
         return TestCase(classname: group.suite, name: group.testCase, time: group.time, skipped: .init(message: nil))
     }
 
     private func generateSuiteStart(line: String) -> String? {
-        guard let _group: any CaptureGroup = line.captureGroup(with: TestSuiteStartCaptureGroup.pattern) else { return nil }
-        guard let group = _group as? TestSuiteStartCaptureGroup else { return nil }
+        let groups = TestSuiteStartCaptureGroup.regex.captureGroups(for: line)
+        guard let group = TestSuiteStartCaptureGroup(groups: groups) else { return nil }
         return group.testSuiteName
     }