--- a/Sources/XcbeautifyLib/Parser.swift
+++ b/Sources/XcbeautifyLib/Parser.swift
@@ -1,13 +1,7 @@
 import Foundation
 
 package class Parser {
-    private let colored: Bool
-
-    private let renderer: any OutputRendering
-
-    private let additionalLines: () -> String?
-
-    private let preserveUnbeautifiedLines: Bool
+    private let formatter: XcbeautifyLib.Formatter
 
     package private(set) var outputType = OutputType.undefined
 
@@ -112,19 +106,12 @@ package class Parser {
         preserveUnbeautifiedLines: Bool = false,
         additionalLines: @escaping () -> (String?)
     ) {
-        self.colored = colored
-
-        switch renderer {
-        case .terminal:
-            self.renderer = TerminalRenderer(colored: colored, additionalLines: additionalLines)
-        case .gitHubActions:
-            self.renderer = GitHubActionsRenderer(colored: colored, additionalLines: additionalLines)
-        case .teamcity:
-            self.renderer = TeamCityRenderer(colored: colored, additionalLines: additionalLines)
-        }
-
-        self.preserveUnbeautifiedLines = preserveUnbeautifiedLines
-        self.additionalLines = additionalLines
+        self.formatter = Formatter(
+            colored: colored,
+            renderer: renderer,
+            preserveUnbeautifiedLines: preserveUnbeautifiedLines,
+            additionalLines: additionalLines
+        )
     }
 
     package func parse(line: String) -> String? {
@@ -139,7 +126,7 @@ package class Parser {
 
             // Nothing found?
             outputType = OutputType.undefined
-            return preserveUnbeautifiedLines ? line : nil
+            return formatter.preserveUnbeautifiedLines ? line : nil
         }
 
         guard let captureGroupType = captureGroupTypes[safe: idx] else {
@@ -153,13 +140,11 @@ package class Parser {
             return nil
         }
 
-        let formattedOutput = renderer.beautify(group: captureGroup)
-
         outputType = captureGroupType.outputType
 
         // Move found parser to the top, so next time it will be checked first
         captureGroupTypes.insert(captureGroupTypes.remove(at: idx), at: 0)
 
-        return formattedOutput
+        return formatter.format(captureGroup: captureGroup)
     }
 }