--- a/Sources/XcbeautifyLib/Regex.swift
+++ b/Sources/XcbeautifyLib/Regex.swift
@@ -2,7 +2,7 @@ import Foundation
 
 // `NSRegularExpression` is marked as `@unchecked Sendable`.
 // Match the definition here.
-final class Regex: @unchecked Sendable {
+package final class Regex: @unchecked Sendable {
     let pattern: String
 
     private lazy var regex: NSRegularExpression? = {