--- a/Sources/XcbeautifyLib/XCBeautifier.swift
+++ b/Sources/XcbeautifyLib/XCBeautifier.swift
@@ -1,7 +1,9 @@
 import Foundation
 
 public struct XCBeautifier {
-    private let parser: Parser
+    private let parser = Parser()
+    private let formatter: Formatter
+    private let preserveUnbeautifiedLines: Bool
 
     public init(
         colored: Bool,
@@ -9,15 +11,24 @@ public struct XCBeautifier {
         preserveUnbeautifiedLines: Bool,
         additionalLines: @escaping () -> String?
     ) {
-        parser = Parser(
+        formatter = Formatter(
             colored: colored,
             renderer: renderer,
-            preserveUnbeautifiedLines: preserveUnbeautifiedLines,
             additionalLines: additionalLines
         )
+
+        self.preserveUnbeautifiedLines = preserveUnbeautifiedLines
     }
 
     public func format(line: String) -> String? {
-        parser.parse(line: line)
+        guard let captureGroup = parser.parse(line: line) else {
+            if preserveUnbeautifiedLines {
+                return line
+            } else {
+                return nil
+            }
+        }
+
+        return formatter.format(captureGroup: captureGroup)
     }
 }