… Sources/XcbeautifyLib/String+CapturedGroups.swift
… Tests/XcbeautifyLibTests/ParsingTests/ParsingTests.swift
… Tests/XcbeautifyLibTests/RendererTests/GitHubActionsRendererTests.swift
… Tests/XcbeautifyLibTests/RendererTests/TeamCityRendererTests.swift
… Tests/XcbeautifyLibTests/RendererTests/TerminalRendererTests.swift