## Changes

### Major

Simplify `Parser` to exclusively implement its intended functionality: parsing. Separate formatting-related logic into a new type, `Formatter`.

Move the `beautify` method from `OutputRendering` to `Formatter`, delete the unnecessary extension on `String`, and update `JunitReporter` as necessary. Update `XCBeautifier` and the executable to wrap `Parser`, `Formatter`, and output-related logic.

### Minor

- Make types `package`-accessible, as necessary.
- Update unit tests to support new implementation.

## Testing

### Performance

| | Old (Seconds) | New (Seconds) | Difference (%) |
|:---:|:---:|:---:|:---:|
| Run 1 | 4.84 | 4.88 | - |
| Run 2 | 4.87 | 4.73 | - |
| Run 3 | 4.82 | 4.79 | - |
| Run 4 | 4.93 | 4.73 | - |
| Run 5 | 4.87 | 4.75 | - |
| Average | 4.866 | 4.776 | -1.87% |